Fix file parsing
Fix the assumption that the structures in size.h
were tightly packed. Define packed size macros
for cases where the on-disk size is needed, and
reserve use of sizeof() for cases where an in-
memory copy is being used. Also fix the symbol
comparison logic for DRI/Alcyon symbols in
size.c

Bump the version in size.c