filefix: Add -n option: excludes hdr size from pad
When padding, filefix assumes there will be an 8k
header prepended to the ROM later by some other
tool, such as romsplit with the '-p' option, or
jagcrypt, so it subtracts 8k from the target file
size when padding up to the requested alignment.
The new '-n' option turns off this behavior. This
may be useful when padding files for use on a
Jaguar CD track or to slot into another ROM as a
subroutine.